+++ acl-v2.1-shim/sources/reader.cpp
@@ -277,7 +277,7 @@ namespace acl_sjson_v21
                 input_tracks = std::move(sjson_track_list.track_list);
                 break;
             default:
-                printf("Unsupported SJSON type: %d\n", sjson_type);
+                printf("Unsupported SJSON type: %d\n", static_cast<int>(sjson_type));
                 return false;
             }
